refactor: update naming case
