+++ packages/react/src/components/list-resource/index.ts
