+++ packages/react/src/theme/component-styles/avatar.ts
