+++ packages/react/src/theme/component-styles/optionItem.ts
