+++ packages/react/src/theme/component-styles/pagination.ts
