+++ packages/react/src/theme/component-styles/slider.ts
