+++ packages/react/src/theme/component-styles/stepper.ts
