+++ packages/react/src/theme/component-styles/tabs.ts
